feat: Controller module
- Request, Response Refactoring
- teamBuildingService's Permission Configurer ( Create / Update / Delete Manager)
