+++ waldreg/application-layer/acceptance/src/test/java/org/waldreg/acceptance/teambuilding/TeamBuildingAcceptanceTest.java
@@ -21,6 +21,13 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.waldreg.acceptance.authentication.AuthenticationAcceptanceTestHelper;
 import org.waldreg.acceptance.user.UserAcceptanceTestHelper;
 import org.waldreg.controller.joiningpool.request.UserRequest;
+import org.waldreg.controller.teambuilding.request.TeamBuildingRequest;
+import org.waldreg.controller.teambuilding.request.TeamBuildingUpdateRequest;
+import org.waldreg.controller.teambuilding.request.TeamRequest;
+import org.waldreg.controller.teambuilding.request.TeamUpdateRequest;
+import org.waldreg.controller.teambuilding.request.UserWeightRequest;
+import org.waldreg.controller.teambuilding.response.TeamBuildingListResponse;
+import org.waldreg.controller.teambuilding.response.TeamBuildingResponse;
 import org.waldreg.controller.user.response.UserResponse;
 
 @SpringBootTest
@@ -1585,490 +1592,6 @@ public class TeamBuildingAcceptanceTest{
             title = title + "A";
         }
         return title;
-    }
-
-    public static class TeamBuildingRequest{
-
-        @JsonProperty("teambuilding_title")
-        private String teamBuildingTitle;
-        @JsonProperty("team_count")
-        private int teamCount;
-        @JsonProperty("users")
-        private List<UserWeightRequest> userList;
-
-        private TeamBuildingRequest(){}
-
-        private TeamBuildingRequest(Builder builder){
-            this.teamBuildingTitle = builder.teamBuildingTitle;
-            this.teamCount = builder.teamCount;
-            this.userList = builder.userList;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public String getTeamBuildingTitle(){
-            return teamBuildingTitle;
-        }
-
-        public int getTeamCount(){
-            return teamCount;
-        }
-
-        public List<UserWeightRequest> getUserList(){
-            return userList;
-        }
-
-        public final static class Builder{
-
-            private String teamBuildingTitle;
-            private int teamCount;
-            private List<UserWeightRequest> userList;
-
-            private Builder(){}
-
-            public Builder teamBuildingTitle(String teamBuildingTitle){
-                this.teamBuildingTitle = teamBuildingTitle;
-                return this;
-            }
-
-            public Builder teamCount(int teamCount){
-                this.teamCount = teamCount;
-                return this;
-            }
-
-            public Builder userList(List<UserWeightRequest> userList){
-                this.userList = userList;
-                return this;
-            }
-
-            public TeamBuildingRequest build(){return new TeamBuildingRequest(this);}
-
-        }
-
-    }
-
-    public static class TeamRequest{
-
-        @JsonProperty("team_name")
-        private String teamName;
-        @JsonProperty("members")
-        private List<String> members;
-
-        private TeamRequest(){}
-
-        private TeamRequest(Builder builder){
-            this.teamName = builder.teamName;
-            this.members = builder.members;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public String getTeamName(){
-            return teamName;
-        }
-
-        public List<String> getMembers(){
-            return members;
-        }
-
-        public final static class Builder{
-
-            private String teamName;
-            private List<String> members;
-
-            public Builder teamName(String teamName){
-                this.teamName = teamName;
-                return this;
-            }
-
-            public Builder members(List<String> members){
-                this.members = members;
-                return this;
-            }
-
-            public TeamRequest build(){return new TeamRequest(this);}
-
-        }
-
-    }
-
-    public static class TeamBuildingUpdateRequest{
-
-        @JsonProperty("teambuilding_title")
-        private String teamBuildingTitle;
-
-        private TeamBuildingUpdateRequest(){}
-
-        private TeamBuildingUpdateRequest(Builder builder){
-            this.teamBuildingTitle = builder.teamBuildingTitle;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public String getTeamBuildingTitle(){
-            return teamBuildingTitle;
-        }
-
-        public final static class Builder{
-
-            private String teamBuildingTitle;
-
-            public Builder teamBuildingTitle(String teamBuildingTitle){
-                this.teamBuildingTitle = teamBuildingTitle;
-                return this;
-            }
-
-            public TeamBuildingUpdateRequest build(){return new TeamBuildingUpdateRequest(this);}
-
-        }
-
-    }
-
-    public static class TeamUpdateRequest{
-
-        @JsonProperty("team_name")
-        private String teamName;
-
-        private TeamUpdateRequest(){}
-
-        private TeamUpdateRequest(Builder builder){
-            this.teamName = builder.teamName;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public String getTeamName(){
-            return teamName;
-        }
-
-        public final static class Builder{
-
-            private String teamName;
-
-            public Builder teamName(String teamName){
-                this.teamName = teamName;
-                return this;
-            }
-
-            public TeamUpdateRequest build(){return new TeamUpdateRequest(this);}
-
-        }
-
-    }
-
-    public static class UserWeightRequest{
-
-        @JsonProperty("user_id")
-        private String userId;
-        @JsonProperty("weight")
-        private int weight;
-
-        private UserWeightRequest(){}
-
-        private UserWeightRequest(Builder builder){
-            this.userId = builder.userId;
-            this.weight = builder.weight;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public String getUserId(){
-            return userId;
-        }
-
-        public int getWeight(){
-            return weight;
-        }
-
-        public void setUserId(String userId){
-            this.userId = userId;
-        }
-
-        public void setWeight(int weight){
-            this.weight = weight;
-        }
-
-        public final static class Builder{
-
-            private String userId;
-            private int weight;
-
-            private Builder(){}
-
-            public Builder userId(String userId){
-                this.userId = userId;
-                return this;
-            }
-
-            public Builder weight(int weight){
-                this.weight = weight;
-                return this;
-            }
-
-            public UserWeightRequest build(){return new UserWeightRequest(this);}
-
-        }
-
-    }
-
-    public static class TeamBuildingListResponse{
-
-        @JsonProperty("max_idx")
-        private int maxIdx;
-        @JsonProperty("teambuildings")
-        private List<TeamBuildingResponse> teamBuildingList;
-
-        private TeamBuildingListResponse(){}
-
-        private TeamBuildingListResponse(Builder builder){
-            this.maxIdx = builder.maxIdx;
-            this.teamBuildingList = builder.teamBuildingList;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public int getMaxIdx(){
-            return maxIdx;
-        }
-
-        public List<TeamBuildingResponse> getTeamBuildingList(){
-            return teamBuildingList;
-        }
-
-        public final static class Builder{
-
-            private int maxIdx;
-            private List<TeamBuildingResponse> teamBuildingList;
-
-            private Builder(){}
-
-            public Builder maxIdx(int maxIdx){
-                this.maxIdx = maxIdx;
-                return this;
-            }
-
-            public Builder teamBuildingList(List<TeamBuildingResponse> teamBuildingList){
-                this.teamBuildingList = teamBuildingList;
-                return this;
-            }
-
-            public TeamBuildingListResponse build(){return new TeamBuildingListResponse(this);}
-
-        }
-
-    }
-
-    public static class TeamBuildingResponse{
-
-        @JsonProperty("teambuilding_id")
-        private int teamBuildingId;
-        @JsonProperty("teambuilding_title")
-        private String teamBuildingTitle;
-        @JsonProperty("created_at")
-        private LocalDateTime createdAt;
-        @JsonProperty("last_modified_at")
-        private LocalDateTime lastModifiedAt;
-        @JsonProperty("teams")
-        private List<TeamResponse> teamList;
-
-        private TeamBuildingResponse(){}
-
-        private TeamBuildingResponse(Builder builder){
-            this.teamBuildingId = builder.teamBuildingId;
-            this.teamBuildingTitle = builder.teamBuildingTitle;
-            this.createdAt = builder.createdAt;
-            this.lastModifiedAt = builder.lastModifiedAt;
-            this.teamList = builder.teamList;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public int getTeamBuildingId(){
-            return teamBuildingId;
-        }
-
-        public String getTeamBuildingTitle(){
-            return teamBuildingTitle;
-        }
-
-        public LocalDateTime getCreatedAt(){
-            return createdAt;
-        }
-
-        public LocalDateTime getLastModifiedAt(){
-            return lastModifiedAt;
-        }
-
-        public List<TeamResponse> getTeamList(){
-            return teamList;
-        }
-
-        public final static class Builder{
-
-            private int teamBuildingId;
-            private String teamBuildingTitle;
-            private LocalDateTime createdAt;
-            private LocalDateTime lastModifiedAt;
-            private List<TeamResponse> teamList;
-
-            public Builder teamBuildingId(int teamBuildingId){
-                this.teamBuildingId = teamBuildingId;
-                return this;
-            }
-
-            public Builder teamBuildingTitle(String teamBuildingTitle){
-                this.teamBuildingTitle = teamBuildingTitle;
-                return this;
-            }
-
-            public Builder createdAt(LocalDateTime createdAt){
-                this.createdAt = createdAt;
-                return this;
-            }
-
-            public Builder lastModifiedAt(LocalDateTime lastModifiedAt){
-                this.lastModifiedAt = lastModifiedAt;
-                return this;
-            }
-
-            public Builder teamList(List<TeamResponse> teamList){
-                this.teamList = teamList;
-                return this;
-            }
-
-            public TeamBuildingResponse build(){return new TeamBuildingResponse(this);}
-
-        }
-
-    }
-
-    public static class TeamResponse{
-
-        @JsonProperty("team_id")
-        private int teamId;
-        @JsonProperty("team_name")
-        private String teamName;
-        @JsonProperty("last_modified_at")
-        private LocalDateTime lastModifiedAt;
-        @JsonProperty("team_members")
-        private List<MemberResponse> memberList;
-
-        private TeamResponse(){}
-
-        private TeamResponse(Builder builder){
-            this.teamId = builder.teamId;
-            this.teamName = builder.teamName;
-            this.lastModifiedAt = builder.lastModifiedAt;
-            this.memberList = builder.memberList;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public int getTeamId(){
-            return teamId;
-        }
-
-        public String getTeamName(){
-            return teamName;
-        }
-
-        public LocalDateTime getLastModifiedAt(){
-            return lastModifiedAt;
-        }
-
-        public List<MemberResponse> getMemberList(){
-            return memberList;
-        }
-
-        public final static class Builder{
-
-            private int teamId;
-            private String teamName;
-            private LocalDateTime lastModifiedAt;
-            private List<MemberResponse> memberList;
-
-            public Builder teamId(int teamId){
-                this.teamId = teamId;
-                return this;
-            }
-
-            public Builder teamName(String teamName){
-                this.teamName = teamName;
-                return this;
-            }
-
-            public Builder lastModified(LocalDateTime lastModifiedAt){
-                this.lastModifiedAt = lastModifiedAt;
-                return this;
-            }
-
-            public Builder memberList(List<MemberResponse> memberList){
-                this.memberList = memberList;
-                return this;
-            }
-
-            public TeamResponse build(){return new TeamResponse(this);}
-
-        }
-
-    }
-
-    public static class MemberResponse{
-
-        @JsonProperty("id")
-        private int id;
-        @JsonProperty("user_id")
-        private String userId;
-        @JsonProperty("name")
-        private String name;
-
-        private MemberResponse(){}
-
-        private MemberResponse(Builder builder){
-            this.id = builder.id;
-            this.userId = builder.userId;
-            this.name = builder.name;
-        }
-
-        public static Builder builder(){return new Builder();}
-
-        public int getId(){
-            return id;
-        }
-
-        public String getUserId(){
-            return userId;
-        }
-
-        public String getName(){
-            return name;
-        }
-
-        public final static class Builder{
-
-            private int id;
-            private String userId;
-            private String name;
-
-            public Builder id(int id){
-                this.id = id;
-                return this;
-            }
-
-            public Builder userId(String userId){
-                this.userId = userId;
-                return this;
-            }
-
-            public Builder name(String name){
-                this.name = name;
-                return this;
-            }
-
-            public MemberResponse build(){return new MemberResponse(this);}
-
-        }
-
     }
 
 }
