+++ waldreg/application-layer/teambuilding-controller/build.gradle
@@ -23,11 +23,18 @@ dependencyManagement {
 }
 
 dependencies {
+    implementation project(':service-layer:teambuilding')
+    implementation project(':repository-layer:memory-repository')
     implementation project(':stable-service-layer:character')
+    implementation project(':core-layer:util')
+    implementation project (':core-layer:template')
+    implementation project(':service-layer:auth')
 
     implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
 
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
