+++ waldreg/application-layer/teambuilding-controller/src/main/java/org/waldreg/controller/teambuilding/TeamBuildingController.java
@@ -1,0 +1,23 @@
+package org.waldreg.controller.teambuilding;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.waldreg.controller.teambuilding.mapper.ControllerTeamBuildingMapper;
+import org.waldreg.teambuilding.teambuilding.management.TeamBuildingManager;
+
+@RestController
+public class TeamBuildingController{
+
+    private final TeamBuildingManager teamBuildingManager;
+
+    private final ControllerTeamBuildingMapper controllerTeamBuildingMapper;
+
+
+    public TeamBuildingController(TeamBuildingManager teamBuildingManager, ControllerTeamBuildingMapper controllerTeamBuildingMapper){
+        this.teamBuildingManager = teamBuildingManager;
+        this.controllerTeamBuildingMapper = controllerTeamBuildingMapper;
+    }
+
+
+
+}
