+++ waldreg/application-layer/teambuilding-controller/src/main/java/org/waldreg/controller/teambuilding/mapper/ControllerTeamBuildingMapper.java
@@ -1,0 +1,4 @@
+package org.waldreg.controller.teambuilding.mapper;
+
+public class ControllerTeamBuildingMapper{
+}
